:bug: fix: reset error message
